requirements.txt
flask>=3.1.0
flask-sqlalchemy>=3.1.1
sqlalchemy>=2.0.36
anthropic>=0.40.0
requests>=2.31.0
apscheduler>=3.10.4
gunicorn>=22.0.0
python-dotenv>=1.0.1
